move common functions to common.go
